+++ grow/settings.py
@@ -41,7 +41,6 @@ INSTALLED_APPS = [
 
 MIDDLEWARE = [
     "django.middleware.security.SecurityMiddleware",
-    "grow.middleware.DeprecationHeadersMiddleware",
     "corsheaders.middleware.CorsMiddleware",
     "the_music_tree_api_kit.view.middleware.HostValidationMiddleware.HostValidationMiddleware",
     "django.middleware.common.CommonMiddleware",